functions/predict/requirements.txt
learns2==0.0.20
tensorflow==2.4.0
google-cloud-storage
google-cloud-firestore